requirements.txt
Flask==2.3.2
Flask-MySQLdb==1.0.1
opencv-python==4.8.0.74
python-dotenv==1.0.0
mysql-connector-python==8.0.33
